requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary
djangorestframework-simplejwt
django-allauth
dj-rest-auth
django-filter
phonenumbers
celery[redis]==5.3.5